requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.0.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dlb==0.6.1
dlib==19.24.6
dm-tree==0.1.8
facenet-pytorch==2.6.0
falcon==4.0.2
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gotrue==2.11.1
grpcio==1.69.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
imutils==0.5.4
ipython-genutils==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
moviepy==2.1.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.1.0
namex==0.0.8
narwhals==1.21.1
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.6
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.2.0
ploomber-cloud==0.3.5
ploomber-core==0.2.26
plotly==5.24.1
postgrest==0.19.1
posthog==3.7.5
proglog==0.1.10
progressbar2==4.5.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-mimeparse==2.0.0
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.1.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
storage3==0.11.0
streamlit==1.41.1
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.15.0
termcolor==2.5.0
tflearn==0.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
